requirements.txt
openai>=1.0.0
python-dotenv>=0.19.0
requests>=2.32.3
urllib3>=2.3.0
charset-normalizer>=3.4.1
idna>=3.10
certifi>=2025.1.31
